requirements.txt
IPython
Pillow >= 3.4.2
absl-py
attrs
backports.tempfile
bokeh >= 0.12.0
Flask
gast == 0.2.2
gym < 0.16.0
imageio
intervaltree >= 2.1.0
Jinja2
joblib >= 0.12
librosa >= 0.6.2
matplotlib >= 1.5.3
mido == 1.2.6
mir_eval >= 0.4
music21
numpy >= 1.14.6
pandas >= 0.18.1
pretty_midi >= 0.2.6
protobuf >= 3.6.1
pygtrie >= 2.3
python-rtmidi >= 1.1, < 1.2
scikit-image
scipy >= 0.18.1
six >= 1.12.0
sk-video
dm-sonnet < 1.36.0
sox >= 1.3.7
tensorflow >= 1.15.0, < 2.0.0
tensorflow-datasets >= 1.0.2
tensorflow-probability == 0.7.0
tensor2tensor >= 1.13.4
wheel
futures;python_version=="2.7"
apache-beam[gcp] >= 2.14.0
avro-python3 !=1.9.2
google-cloud-storage
google-cloud-datastore
google-api-python-client
Werkzeug
flask-login
pyjwt==1.7.1
flask-cors==3.0.8
google-auth
requests==2.23.0
requests-toolbelt==0.9.1